Fixed some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/2000



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@9740 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/common/ucnvbocu.c
@@ -318,7 +318,7 @@ packDiff(int32_t diff) {
              */
             result|=BOCU1_TRAIL_TO_BYTE(diff)<<16;
 
-            result|=BOCU1_START_POS_4<<24;
+            result|=((uint32_t)BOCU1_START_POS_4)<<24;
         }
     } else {
         /* two- to four-byte negative differences */
